services/rag-service/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
qdrant-client 